Delete Div with list of sites. Now one attribute variables "sites" is use for listing searched site or listing all site in function of the sent method (post or get).
Added an attribute variable "list" with boolean value. If variables "sites" contains searched sites, "list" is false, and messages corresponding at searched sites is displayed. If variables "sites" contains the list of all blog's sites, "list" is true, and messages corresponding at list of all sites is displayed.

Modify some things in the search banner. For each select, added an submit event "on change" in order to launch the search with each change in the select.

In tag option of select tag, add a test to add the attribute selected at the selected option.